# doc/openapi.yaml
openapi: 3.0.0
info:
  title: Myelin Base RAG API
  version: 1.0.0
  description: API documentation for Myelin Base RAG backend
servers:
  - url: /api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - INVALID_PARAMETER
        - MISSING_PARAMETER
        - PERMISSION_DENIED
        - RESOURCE_NOT_FOUND
        - INTERNAL_SERVER_ERROR
        - DOCUMENTS_FILE_TOO_LARGE
        - DOCUMENTS_UNSUPPORTED_FILE_TYPE
        - DOCUMENTS_INVALID_FILENAME_LENGTH_LIMIT
        - DOCUMENTS_TAG_LENGTH_LIMIT
        - DOCUMENTS_TAGS_TOO_MANY
        - DOCUMENTS_NOT_FOUND
        - DOCUMENTS_UPLOAD_FAILED
        - DOCUMENTS_NOT_READY_FOR_DOWNLOAD
        - DOCUMENTS_FILENAME_EMPTY
        - DOCUMENTS_SELECTION_EMPTY
        - DOCUMENTS_SELECTION_TOO_MANY
        - DOCUMENTS_DUPLICATE_CONTENT
        - CHAT_SESSION_NAME_TOO_LONG
        - CHAT_SESSION_NAME_EMPTY
        - CHAT_QUERY_EMPTY
        - CHAT_QUERY_TOO_LONG
        - CHAT_COMMENT_TOO_LONG
        - CHAT_FEEDBACK_REASONS_INVALID
        - CHAT_FEEDBACK_REASONS_EMPTY
    ErrorResponse:
      type: object
      properties:
        errorCode:
          $ref: "#/components/schemas/ErrorCode"
      required:
        - errorCode
...
    DocumentStatus:
      type: string
      enum:
        - PENDING_UPLOAD
        - PROCESSING
        - COMPLETED
        - FAILED
        - DELETED
        - DELETE_FAILED
    DocumentResponse:
      type: object
      properties:
        documentId:
          type: string
        fileName:
          type: string
        contentType:
          type: string
        fileSize:
          type: number
        tags:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/DocumentStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tagUpdatedAt:
          type: string
          format: date-time
        processingStatus:
          type: string
        errorMessage:
          type: string
        downloadUrl:
          type: string
      required:
        - documentId
        - fileName
        - contentType
        - fileSize
        - tags
        - status
        - createdAt
        - updatedAt
    FileMetadata:
      type: object
      properties:
        fileName:
          type: string
          minLength: 1
          maxLength: 255
        contentType:
          type: string
        fileSize:
          type: number
          minimum: 0
          maximum: 52428800
        fileHash:
          type: string
          minLength: 64
          maxLength: 64
      required:
        - fileName
        - contentType
        - fileSize
    UploadRequestRequest:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileMetadata"
          minItems: 1
          maxItems: 100
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
          default: []
      required:
        - files
    UpdateTagsRequest:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
      required:
        - tags
    BatchDeleteRequest:
      type: object
      properties:
        documentIds:
          type: array
          items:
            type: string
            maxLength: 100
          minItems: 1
          maxItems: 100
      required:
        - documentIds
    GetDocumentsResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentResponse"
      required:
        - documents
    GetDocumentResponse:
      type: object
      properties:
        document:
          $ref: "#/components/schemas/DocumentResponse"
      required:
        - document
    GetDocumentDownloadUrlResponse:
      type: object
      properties:
        downloadUrl:
          type: string
      required:
        - downloadUrl
    UploadRequestResult:
      type: object
      properties:
        documentId:
          type: string
        fileName:
          type: string
        uploadUrl:
          type: string
        expiresIn:
          type: number
        s3Key:
          type: string
      required:
        - documentId
        - fileName
        - uploadUrl
        - expiresIn
        - s3Key
    UploadRequestFileResult:
      type: object
      properties:
        status:
          type: string
          enum:
            - success
            - error
        fileName:
          type: string
        data:
          $ref: "#/components/schemas/UploadRequestResult"
        errorCode:
          type: string
      required:
        - status
        - fileName
    UploadRequestResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/UploadRequestFileResult"
      required:
        - results
    DeleteDocumentResponse:
      type: object
      properties:
        document:
          $ref: "#/components/schemas/DocumentResponse"
      required:
        - document
    ResultStatus:
      type: string
      enum:
        - success
        - error
    BatchDeleteResult:
      type: object
      properties:
        documentId:
          type: string
        status:
          $ref: "#/components/schemas/ResultStatus"
        errorCode:
          type: string
      required:
        - documentId
        - status
    BatchDeleteResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/BatchDeleteResult"
      required:
        - results
    UploadRequestErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          enum:
            - DOCUMENTS_SELECTION_EMPTY
            - DOCUMENTS_TAG_LENGTH_LIMIT
            - DOCUMENTS_TAGS_TOO_MANY
            - DOCUMENTS_FILE_TOO_LARGE
            - DOCUMENTS_INVALID_FILENAME_LENGTH_LIMIT
            - DOCUMENTS_UNSUPPORTED_FILE_TYPE
            - DOCUMENTS_FILENAME_EMPTY
            - DOCUMENTS_DUPLICATE_CONTENT
            - VALIDATION_FAILED
            - INVALID_PARAMETER
      required:
        - errorCode
    UpdateTagsErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          enum:
            - MISSING_PARAMETER
            - DOCUMENTS_TAG_LENGTH_LIMIT
            - DOCUMENTS_TAGS_TOO_MANY
            - RESOURCE_NOT_FOUND
            - PERMISSION_DENIED
            - VALIDATION_FAILED
            - INVALID_PARAMETER
      required:
        - errorCode
    BatchDeleteErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          enum:
            - DOCUMENTS_SELECTION_EMPTY
            - VALIDATION_FAILED
            - INVALID_PARAMETER
      required:
        - errorCode
    GetDocumentErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          enum:
            - MISSING_PARAMETER
            - DOCUMENTS_NOT_FOUND
            - PERMISSION_DENIED
      required:
        - errorCode
    GetDocumentDownloadUrlErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          enum:
            - MISSING_PARAMETER
            - DOCUMENTS_NOT_FOUND
            - PERMISSION_DENIED
            - DOCUMENTS_NOT_READY_FOR_DOWNLOAD
      required:
        - errorCode
    DeleteDocumentErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          enum:
            - MISSING_PARAMETER
            - RESOURCE_NOT_FOUND
            - PERMISSION_DENIED
      required:
        - errorCode
...
paths:
...
  /documents:
    get:
      summary: List Documents
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDocumentsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/upload:
    post:
      summary: Request Upload URL
      description: S3への署名付きアップロードURLを発行します。
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequestRequest"
      responses:
        "202":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadRequestResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadRequestErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/batch-delete:
    post:
      summary: Batch Delete Documents
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchDeleteRequest"
      responses:
        "200":
          description: Batch delete completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchDeleteResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchDeleteErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}:
    get:
      summary: Get Document
      security:
        - BearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDocumentResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDocumentErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDocumentErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete Document
      security:
        - BearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "202":
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteDocumentResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteDocumentErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteDocumentErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/download-url:
    get:
      summary: Get Document Download URL
      security:
        - BearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDocumentDownloadUrlResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDocumentDownloadUrlErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDocumentDownloadUrlErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/tags:
    patch:
      summary: Update Tags
      security:
        - BearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTagsRequest"
      responses:
        "204":
          description: Tags updated successfully
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateTagsErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateTagsErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"