# infrastructure/src/functions/documents/index.ts
import {
  GetCommand,
  PutCommand,
  QueryCommand,
  UpdateCommand,
} from "@aws-sdk/lib-dynamodb";
import { APIGatewayProxyEvent } from "aws-lambda";
import { randomUUID } from "crypto";

import {
  createPineconeClient,
  deleteDocumentVectors,
  getPineconeApiKey,
} from "../../shared/clients/pinecone";
import {
  BatchDeleteRequestDto,
  BatchDeleteRequestSchema,
  BatchDeleteResponseDto,
  BatchDeleteResultDto,
  FileMetadataDto,
  GetDocumentResponseDto,
  GetDocumentsResponseDto,
  ResultStatusSchema,
  TEXT_TYPES,
  UpdateTagsRequestDto,
  UpdateTagsRequestSchema,
  UploadRequestFileResultDto,
  UploadRequestRequestDto,
  UploadRequestRequestSchema,
  UploadRequestResponseDto,
  UploadRequestResultDto,
} from "../../shared/schemas/dto/document.dto";
import {
  DocumentEntity,
  DocumentStatusSchema,
} from "../../shared/schemas/entities/document.entity";
import { ErrorCode } from "../../shared/types/error-code";
import {
  apiHandler,
  AppError,
  logger,
  validateJson,
} from "../../shared/utils/api-handler";
import { toDocumentDTO } from "../../shared/utils/dto-mapper";
import { createDynamoDBClient } from "../../shared/utils/dynamodb";
import {
  buildS3Uri,
  createS3Client,
  generateDownloadUrl,
  generateUploadUrl,
} from "../../shared/utils/s3";

const TABLE_NAME = process.env.TABLE_NAME!;
const BUCKET_NAME = process.env.BUCKET_NAME!;
const PRESIGNED_URL_EXPIRY = parseInt(
  process.env.PRESIGNED_URL_EXPIRY || "900",
  10
);
const IS_LOCAL_STAGE = process.env.STAGE! === "local";
const DOCUMENT_TTL_SECONDS = 24 * 60 * 60;

const docClient = createDynamoDBClient();
const s3Client = createS3Client();

export const handler = apiHandler(async (event: APIGatewayProxyEvent) => {
  const { httpMethod, path, pathParameters } = event;
  const ownerId = extractOwnerId(event);

  if (httpMethod === "GET" && path === "/documents") {
    const response = await getDocuments(ownerId);
    return {
      body: response,
    };
  }

  if (
    httpMethod === "GET" &&
    path.startsWith("/documents/") &&
    path.endsWith("/download-url")
  ) {
    const documentId = pathParameters?.id;
    if (!documentId) throw new AppError(400, ErrorCode.MISSING_PARAMETER);

    const response = await getDownloadUrl(documentId, ownerId);
    return {
      body: response,
    };
  }

  if (httpMethod === "GET" && path.startsWith("/documents/")) {
    const documentId = pathParameters?.id;
    if (!documentId) throw new AppError(400, ErrorCode.MISSING_PARAMETER);

    const response = await getDocumentById(documentId, ownerId);
    return {
      body: response,
    };
  }

  if (httpMethod === "POST" && path === "/documents/upload") {
    const body = validateJson(event.body, UploadRequestRequestSchema);
    const response = await uploadRequest(body, ownerId);
    return { statusCode: 202, body: response };
  }

  if (httpMethod === "DELETE" && path.startsWith("/documents/")) {
    const documentId = pathParameters?.id;
    if (!documentId) throw new AppError(400, ErrorCode.MISSING_PARAMETER);

    await deleteDocument(documentId, ownerId);
    return { statusCode: 202 };
  }

  if (httpMethod === "POST" && path === "/documents/batch-delete") {
    const body = validateJson(event.body, BatchDeleteRequestSchema);
    const response = await batchDeleteDocuments(body, ownerId);
    return { statusCode: 200, body: response };
  }

  if (httpMethod === "PATCH" && path.endsWith("/tags")) {
    const documentId = pathParameters?.id;
    if (!documentId) throw new AppError(400, ErrorCode.MISSING_PARAMETER);

    const body = validateJson(event.body, UpdateTagsRequestSchema);
    await updateTags(documentId, body, ownerId);
    return { statusCode: 204 };
  }

  throw new AppError(404);
});

function extractOwnerId(event: APIGatewayProxyEvent): string {
  if (IS_LOCAL_STAGE) {
    return "user-001";
  }

  const claims = event.requestContext?.authorizer?.claims;
  const ownerId = claims?.sub;

  if (!ownerId) throw new AppError(401, ErrorCode.PERMISSION_DENIED);

  return ownerId;
}

/**
 * ドキュメント一覧取得 GET /documents
 */
async function getDocuments(ownerId: string): Promise<GetDocumentsResponseDto> {
  const command = new QueryCommand({
    TableName: TABLE_NAME,
    IndexName: "OwnerIndex",
    FilterExpression: "#status <> :deleted",
    KeyConditionExpression: "#owner = :ownerId",
    ExpressionAttributeNames: { "#owner": "ownerId", "#status": "status" },
    ExpressionAttributeValues: {
      ":ownerId": ownerId,
      ":deleted": DocumentStatusSchema.enum.DELETED,
    },
    ScanIndexForward: false,
  });

  const response = await docClient.send(command);
  const items = (response.Items || []) as DocumentEntity[];

  const documents = items.map((item) => toDocumentDTO(item));
  return { documents };
}

/**
 * ダウンロードURL取得 GET /documents/{documentId}/download-url
 */
async function getDownloadUrl(
  documentId: string,
  ownerId: string
): Promise<{ downloadUrl: string }> {
  const command = new GetCommand({
    TableName: TABLE_NAME,
    Key: { documentId },
  });

  const response = await docClient.send(command);
  const item = response.Item as DocumentEntity;

  if (!item) throw new AppError(404, ErrorCode.DOCUMENTS_NOT_FOUND);

  if (item.ownerId !== ownerId) {
    logger("WARN", "Access attempt to document by non-owner", {
      documentId,
      ownerId,
      actualOwnerId: item.ownerId,
    });
    // セキュリティ上の理由で404を返すが、エラーコードでPERMISSION_DENIEDを明示
    throw new AppError(404, ErrorCode.PERMISSION_DENIED);
  }

  if (item.status === DocumentStatusSchema.enum.DELETED) {
    throw new AppError(404, ErrorCode.DOCUMENTS_NOT_FOUND);
  }

  if (item.status !== DocumentStatusSchema.enum.COMPLETED || !item.s3Key) {
    throw new AppError(400, ErrorCode.DOCUMENTS_NOT_READY_FOR_DOWNLOAD);
  }

  let responseContentType = item.contentType;

  if (TEXT_TYPES.includes(item.contentType)) {
    if (!responseContentType.includes("charset")) {
      responseContentType = `${responseContentType}; charset=utf-8`;
    }
  }

  const downloadUrl = await generateDownloadUrl(
    s3Client,
    BUCKET_NAME,
    item.s3Key,
    responseContentType,
    3600
  );
  return { downloadUrl };
}

/**
 * ドキュメント詳細取得 GET /documents/{documentId}
 */
async function getDocumentById(
  documentId: string,
  ownerId: string
): Promise<GetDocumentResponseDto> {
  const command = new GetCommand({
    TableName: TABLE_NAME,
    Key: { documentId },
  });

  const response = await docClient.send(command);
  const item = response.Item as DocumentEntity;

  if (!item) throw new AppError(404, ErrorCode.DOCUMENTS_NOT_FOUND);

  if (item.ownerId !== ownerId) {
    logger("WARN", "Access attempt to document by non-owner", {
      documentId,
      ownerId,
      actualOwnerId: item.ownerId,
    });
    // セキュリティ上の理由で404を返すが、エラーコードでPERMISSION_DENIEDを明示
    throw new AppError(404, ErrorCode.PERMISSION_DENIED);
  }

  if (item.status === DocumentStatusSchema.enum.DELETED) {
    throw new AppError(404, ErrorCode.DOCUMENTS_NOT_FOUND);
  }

  const document = toDocumentDTO(item);

  return { document };
}

/**
 * ドキュメント削除 (論理削除 + TTL) DELETE /documents/{documentId}
 */
async function deleteDocument(
  documentId: string,
  ownerId: string
): Promise<void> {
  const ttl = Math.floor(Date.now() / 1000) + DOCUMENT_TTL_SECONDS;

  const command = new UpdateCommand({
    TableName: TABLE_NAME,
    Key: { documentId },
    ConditionExpression: "attribute_exists(documentId) AND ownerId = :ownerId",
    UpdateExpression:
      "SET #status = :deleted, #ttl = :ttl, updatedAt = :updatedAt, processingStatus = :active",
    ExpressionAttributeNames: { "#status": "status", "#ttl": "ttl" },
    ExpressionAttributeValues: {
      ":deleted": DocumentStatusSchema.enum.DELETED,
      ":ttl": ttl,
      ":updatedAt": new Date().toISOString(),
      ":ownerId": ownerId,
      ":active": "ACTIVE",
    },
  });

  await docClient.send(command);

  await deletePineconeVectorsImmediately(documentId);
}

/**
 * ドキュメント一括削除 POST /documents/batch-delete
 */
async function batchDeleteDocuments(
  body: BatchDeleteRequestDto,
  ownerId: string
): Promise<BatchDeleteResponseDto> {
  const { documentIds } = body;

  const results: BatchDeleteResultDto[] = await Promise.all(
    documentIds.map(async (documentId) => {
      try {
        await deleteDocument(documentId, ownerId);
        return {
          documentId,
          status: ResultStatusSchema.enum.success,
        };
      } catch (error) {
        logger("ERROR", "Failed to delete document in batch", {
          documentId,
          ownerId,
          error,
        });
        return {
          documentId,
          status: ResultStatusSchema.enum.error,
          errorCode: ErrorCode.INTERNAL_SERVER_ERROR,
        };
      }
    })
  );

  return { results };
}

/**
 * ドキュメントタグ更新 PATCH /documents/{documentId}/tags
 */
async function updateTags(
  documentId: string,
  body: UpdateTagsRequestDto,
  ownerId: string
): Promise<void> {
  const { tags } = body;

  const sanitizedTags = sanitizeTags(tags);

  const command = new UpdateCommand({
    TableName: TABLE_NAME,
    Key: { documentId },
    ConditionExpression: "attribute_exists(documentId) AND ownerId = :ownerId",
    UpdateExpression:
      "SET #tags = :tags, tagUpdatedAt = :updatedAt, updatedAt = :updatedAt",
    ExpressionAttributeNames: { "#tags": "tags" },
    ExpressionAttributeValues: {
      ":tags": sanitizedTags,
      ":updatedAt": new Date().toISOString(),
      ":ownerId": ownerId,
    },
  });

  await docClient.send(command);
}

/**
 * ドキュメントアップロードリクエスト POST /documents/upload
 */
async function uploadRequest(
  body: UploadRequestRequestDto,
  ownerId: string
): Promise<UploadRequestResponseDto> {
  const { files, tags } = body;

  const sanitizedTags = sanitizeTags(tags);

  const results: UploadRequestFileResultDto[] = await Promise.all(
    files.map(async (file: FileMetadataDto) => {
      try {
        if (file.fileHash) {
          const duplicateByHash = await checkDuplicateByHash(
            ownerId,
            file.fileHash
          );
          if (duplicateByHash) {
            return {
              status: ResultStatusSchema.enum.error,
              fileName: file.fileName,
              errorCode: ErrorCode.DOCUMENTS_DUPLICATE_CONTENT,
            };
          }
        }

        const duplicateDocs = await docClient.send(
          new QueryCommand({
            TableName: TABLE_NAME,
            IndexName: "FileNameIndex",
            KeyConditionExpression:
              "ownerId = :ownerId AND fileName = :fileName",
            ExpressionAttributeValues: {
              ":ownerId": ownerId,
              ":fileName": file.fileName,
              ":deleted": DocumentStatusSchema.enum.DELETED,
            },
            FilterExpression: "#status <> :deleted",
            ExpressionAttributeNames: { "#status": "status" },
          })
        );

        if (duplicateDocs.Items && duplicateDocs.Items.length > 0) {
          const activeDocs = duplicateDocs.Items as DocumentEntity[];

          await Promise.all(
            activeDocs.map((oldDoc) =>
              markDocumentForDeletion(oldDoc.documentId, ownerId)
            )
          );
        }

        const uploadData = await createUploadRequest(
          file,
          sanitizedTags,
          ownerId
        );

        return {
          status: ResultStatusSchema.enum.success,
          fileName: file.fileName,
          data: uploadData,
        };
      } catch (error) {
        logger("ERROR", "Failed to process document upload", {
          fileName: file.fileName,
          fileSize: file.fileSize,
          ownerId: ownerId,
          error: error,
        });
        return {
          status: ResultStatusSchema.enum.error,
          fileName: file.fileName,
          errorCode: ErrorCode.DOCUMENTS_UPLOAD_FAILED,
        };
      }
    })
  );

  return { results };
}

/**
 * コンテンツハッシュによる重複ドキュメントの検出
 * 同一ユーザーが同じ内容のファイルを既にアップロード済みかをチェック
 */
async function checkDuplicateByHash(
  ownerId: string,
  fileHash: string
): Promise<string | null> {
  const result = await docClient.send(
    new QueryCommand({
      TableName: TABLE_NAME,
      IndexName: "FileHashIndex",
      KeyConditionExpression: "ownerId = :ownerId AND fileHash = :fileHash",
      ExpressionAttributeValues: {
        ":ownerId": ownerId,
        ":fileHash": fileHash,
        ":deleted": DocumentStatusSchema.enum.DELETED,
      },
      FilterExpression: "#status <> :deleted",
      ExpressionAttributeNames: { "#status": "status" },
      Limit: 1,
    })
  );

  if (result.Items && result.Items.length > 0) {
    return result.Items[0].documentId;
  }

  return null;
}

/**
 * ドキュメントを削除リクエストにする (同名ファイルアップロード時の置換用)
 */
async function markDocumentForDeletion(documentId: string, ownerId: string) {
  const ttl = Math.floor(Date.now() / 1000) + DOCUMENT_TTL_SECONDS;

  await docClient.send(
    new UpdateCommand({
      TableName: TABLE_NAME,
      Key: { documentId },
      ConditionExpression: "ownerId = :ownerId",
      UpdateExpression:
        "SET #status = :deleted, #ttl = :ttl, updatedAt = :now, processingStatus = :active",
      ExpressionAttributeNames: { "#status": "status", "#ttl": "ttl" },
      ExpressionAttributeValues: {
        ":deleted": DocumentStatusSchema.enum.DELETED,
        ":ttl": ttl,
        ":now": new Date().toISOString(),
        ":ownerId": ownerId,
        ":active": "ACTIVE",
      },
    })
  );

  await deletePineconeVectorsImmediately(documentId);
}

/**
 * ドキュメントアップロードリクエストを作成
 */
async function createUploadRequest(
  file: FileMetadataDto,
  tags: string[],
  ownerId: string
): Promise<UploadRequestResultDto> {
  const documentId = randomUUID();
  const s3Key = `uploads/${ownerId}/${documentId}`;
  const s3Path = buildS3Uri(BUCKET_NAME, s3Key);
  const now = new Date().toISOString();

  const item: DocumentEntity = {
    documentId,
    status: DocumentStatusSchema.enum.PENDING_UPLOAD,
    processingStatus: "ACTIVE",
    fileName: file.fileName,
    contentType: file.contentType,
    fileSize: file.fileSize,
    s3Path,
    s3Key,
    ownerId: ownerId,
    createdAt: now,
    updatedAt: now,
    tags: tags,
    ...(file.fileHash && { fileHash: file.fileHash }),
  };

  await docClient.send(new PutCommand({ TableName: TABLE_NAME, Item: item }));

  const uploadUrl = await generateUploadUrl(
    s3Client,
    BUCKET_NAME,
    s3Key,
    file.contentType,
    PRESIGNED_URL_EXPIRY
  );

  return {
    documentId,
    fileName: file.fileName,
    uploadUrl,
    expiresIn: PRESIGNED_URL_EXPIRY,
    s3Key,
  };
}

/**
 * Pineconeからベクトルを即座に削除
 */
async function deletePineconeVectorsImmediately(
  documentId: string
): Promise<void> {
  if (IS_LOCAL_STAGE) {
    return;
  }

  try {
    const apiKey = await getPineconeApiKey();
    const pinecone = createPineconeClient(apiKey);
    await deleteDocumentVectors(pinecone, documentId);
  } catch (error) {
    logger("ERROR", "Failed to delete Pinecone vectors immediately", {
      documentId,
      error: error instanceof Error ? error.message : String(error),
    });
    throw error;
  }
}

/**
 * タグを正規化
 */
function sanitizeTags(tags: string[]): string[] {
  return tags.map((t) => t.trim()).filter((t) => t.length > 0);
}